refactor(utilities): move is_async_fn from asyncutils to _utilities for better module organization chore: delete asyncutils.py as it is no longer needed
